Add files via upload
1. JAVA 
A. Create an array with the values (1, 2, 3, 4, 5, 6, 7) and shuffle it. 
B. Enter a Roman Number as input and convert it to an integer. (ex IX = 9) 
C. Check if the input is pangram or not. (Pangram is a sentence that contains all the alphabet from a-z) 
2 JavaScript 
A. Take a sentence as an input and reverse every word in that sentence.  
a. Example - This is a sunny day > shiT si a ynnus yad. 
B. Perform sorting of an array in descending order. 
3. HTML 
A. Create a basic calculator using HTML, CSS, and JavaScript with the functionality of add, subtract, multiply and divide. Use the following picture for reference. 
B. Create a survey form with Fields; First Name, Last Name, Date of Birth, Country (dropdown), Gender (checkbox),  Profession,  email,  and  mobile  number.  All  the  input  fields  are necessary to submit the form. Create two buttons Submit and Reset. Reset will reset the form while clicking on submit, first, it will check all the fields and necessary validations and then a popup will appear displaying all the selected values with the label in front of it. On closing the popup, the form should reset all the values. Use the following for reference